Extract coordination forwarder
